BEEPER wip

+++ include/BeeperWaveformGenerator.h
@@ -18,7 +18,7 @@ public:
 
 private:
   bool _state = false;
-  uint32_t _countChangeOrSample = 0;
+  int64_t _countChangeOrSample = 0;
   int _accumZero = 0;
   int _accumOne = 0;
 };
